chore: test 폴더 생성 및 utils 분리
