+++ @wabinar/crdt/test/convergence.test.ts
@@ -1,42 +1,11 @@
-import CRDT from './index';
-import LinkedList, { RemoteInsertOperation } from './linked-list';
-import { Node } from './node';
-
-/**
- * utilities
- * 바로 전달하면 같은 인스턴스를 가리키게 되어 remote operation 의미가 사라짐
- */
-const deepCopyRemoteInsertion = (op: RemoteInsertOperation) => {
-  const { node } = op;
-
-  const copy = { ...node };
-  Object.setPrototypeOf(copy, Node.prototype);
-
-  return { node: copy as Node };
-};
-
-const remoteInsertThroughSocket = (crdt: CRDT, op: RemoteInsertOperation) => {
-  const copy = deepCopyRemoteInsertion(op);
-  crdt.remoteInsert(copy);
-};
-
-/**
- * 모든 remote site 문자열이 일치하는지 확인
- */
-const convergenceCheck = () => {
-  expect(원희.read()).toEqual(주영.read());
-  expect(주영.read()).toEqual(도훈.read());
-  expect(도훈.read()).toEqual(세영.read());
-};
-
-/**
- * 테스트용 site들
- * 인스턴스를 재사용하기 때문에 테스트 케이스 간 의존성 존재
- */
-const 원희 = new CRDT(1, new LinkedList());
-const 주영 = new CRDT(2, new LinkedList());
-const 도훈 = new CRDT(3, new LinkedList());
-const 세영 = new CRDT(4, new LinkedList());
+import {
+  원희,
+  주영,
+  도훈,
+  세영,
+  convergenceCheck,
+  remoteInsertThroughSocket,
+} from './utils';
 
 describe('Convergence', () => {
   it('하나의 site에서 삽입', () => {
